+++ packages/google_maps_flutter/google_maps_flutter_android/CHANGELOG.md
@@ -1,5 +1,10 @@
-## NEXT
+## 2.19.10
 
+* Updates README to include setup information.
+
+## 2.19.9
+
+* Adds attribution ID for Google Maps SDK usage.
 * Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
 
 ## 2.19.8